+++ AccelByte.Sdk/Api/Eventlog/Operation/EventV2/GetEventSpecificUserV2Handler.cs
@@ -29,9 +29,9 @@ namespace AccelByte.Sdk.Api.Eventlog.Operation
             
             public string? EventName { get; set; }
             
-            public double? Offset { get; set; }
+            public long? Offset { get; set; }
             
-            public double? PageSize { get; set; }
+            public long? PageSize { get; set; }
             
             public string? StartDate { get; set; }
             
@@ -50,13 +50,13 @@ namespace AccelByte.Sdk.Api.Eventlog.Operation
                 return this;
             }
 
-            public GetEventSpecificUserV2HandlerBuilder SetOffset(double _offset)
+            public GetEventSpecificUserV2HandlerBuilder SetOffset(long _offset)
             {
                 Offset = _offset;
                 return this;
             }
 
-            public GetEventSpecificUserV2HandlerBuilder SetPageSize(double _pageSize)
+            public GetEventSpecificUserV2HandlerBuilder SetPageSize(long _pageSize)
             {
                 PageSize = _pageSize;
                 return this;
@@ -108,8 +108,8 @@ namespace AccelByte.Sdk.Api.Eventlog.Operation
             string userId,            
             string? endDate,            
             string? eventName,            
-            double? offset,            
-            double? pageSize,            
+            long? offset,            
+            long? pageSize,            
             string? startDate            
         )
         {
